Fixed an issue with program hanging.
We shouldn't ever wait for the V-Blank when the LCD is off as it will never reach V-Blank and therefore execution will hang. We could add checks for this in the waitvblank sub-routine or simply turn the LCD on at the start of waitvblank but I don't like either of these and think we should simply be strict and careful of our usage.